requirements.txt
torch
torchaudio
numpy
sounddevice
faster-whisper
anthropic
python-dotenv
pyttsx3
ollama
mcp
httpx
tavily-python
